+++ Tests/Editor/AnalyzeRules/AnalyzeRuleBaseTests.cs
@@ -22,6 +22,7 @@ namespace UnityEditor.AddressableAssets.Tests.AnalyzeRules
         {
             var bundleName = "2398471298347129034_bundlename_1";
             var fakeFileName = "archive://3912983hf9sdf902340jidf";
+            var convertedBundleName = "group1_bundlename_1";
 
             var group = Settings.CreateGroup("group1", false, false, false, null, typeof(BundledAssetGroupSchema));
 
@@ -38,10 +39,15 @@ namespace UnityEditor.AddressableAssets.Tests.AnalyzeRules
             var field = typeof(ExtractDataTask).GetField("m_WriteData", BindingFlags.NonPublic | BindingFlags.Instance);
             field.SetValue(baseRule.m_ExtractData, new BundleWriteData());
 
+            baseRule.m_AllBundleInputDefs.Add(new AssetBundleBuild()
+            {
+                assetBundleName = bundleName
+            });
+
             baseRule.m_ExtractData.WriteData.FileToBundle.Add(fakeFileName, bundleName);
             baseRule.ConvertBundleNamesToGroupNames(context);
 
-            Assert.AreEqual(group.Name, baseRule.m_ExtractData.WriteData.FileToBundle[fakeFileName]);
+            Assert.AreEqual(convertedBundleName, baseRule.m_ExtractData.WriteData.FileToBundle[fakeFileName]);
 
             Settings.RemoveGroup(group);
         }